Added an optional epsilon to line drawing.
Nudges the line in one direction or another so all the on-the-line samples are pushed in the same direction for that line.

Also, Ash, god damn figuring out the right magic file structure to get tests to work took SO much longer than it would have done to just not write tests, I swear to god.